maui/issues/20046 오류 수정. (CollectionView로 재구현)

+++ CGVModClient/ViewModels/OpenNotificationSettingViewModel.cs
@@ -4,15 +4,15 @@ using Android.Content;
 using CGVModClient.Platforms.Android.Services;
 #endif
 using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
+using System.Collections.ObjectModel;
 
 namespace CGVModClient.ViewModels;
 
 public partial class OpenNotificationSettingViewModel : ObservableObject, IViewModel
 {
     private AppDatabase _database;
-
-    [ObservableProperty]
-    private List<OpenNotificationInfo> infos;
+    public ObservableCollection<OpenNotificationInfo> Infos { get; private set; }
     public bool IsOpenNotificationEnabled {
         get => Preferences.Get("IsOpenNotificationEnabled", false);
         set { 
@@ -24,6 +24,7 @@ public partial class OpenNotificationSettingViewModel : ObservableObject, IViewM
     public OpenNotificationSettingViewModel(AppDatabase database)
     {
         _database = database;
+        Infos = new ObservableCollection<OpenNotificationInfo>();
     }
 
     public async Task LoadAsync()
@@ -40,13 +41,23 @@ public partial class OpenNotificationSettingViewModel : ObservableObject, IViewM
             }
         }
 #endif
-        Infos = await _database.GetOpenNotificationInfosAsync();
+        var infos = await _database.GetOpenNotificationInfosAsync();
+        var comparer = EqualityComparer<OpenNotificationInfo>.Create((x, y) => x?.Id == y?.Id);
+        for (var i = 0; i < infos.Count; i++)
+        {
+            if (!Infos.Contains(infos[i], comparer))
+                Infos.Add(infos[i]);
+        }
     }
 
+    [RelayCommand]
     public async Task RemoveOpenNotificationInfo(OpenNotificationInfo info)
     {
+        var result = await Microsoft.Maui.Controls.Application.Current.MainPage
+            .DisplayAlert("예매 오픈 알림 설정", "알림을 삭제하시겠습니까?", "삭제", "취소", FlowDirection.LeftToRight);
+        if (!result)
+            return;
         Infos.Remove(info);
-        OnPropertyChanged(nameof(Infos));
         await _database.DeleteOpenNotificationInfo(info);
     }
 }
